// src/app/model/advisor.ts
export class advisor {
    advisorId: number;
    portfolioId: number;
    username: string;
    advisorName: string;
    macd: boolean;
    sma: boolean;
    ema: boolean;
    wma: boolean;
    rsi: boolean;
    stochrsi: boolean;
    stoch: boolean;
    willr: boolean;
    ultosc: boolean;
    adxr: boolean;
    mom: boolean;
    cci: boolean;
}